Having a look at input type=color.
So Chrome is weird; the colour part of the element has a reduced height
compared to what you set it to. This makes me want to use a dummy UI
element that calls the picker, or use farbtastic like I do for my
planner website.
It's as I would expect on the current Firefox nightly (with it enabled
in about:config), though it doesn't get clipped by border-radius + the
picker comes up in the top left always. But it just landed so I'm not
concerned.